+++ dashboard/crud/annee/annee.php
@@ -1,22 +1,8 @@
 <?php
 session_start();
 
-require_once('../../../librairies/database/database.php');
-require_once('../../../librairies/patron.php');
-require_once('../../../librairies/models/Annee.php');
+require_once('../../../librairies/controllers/Annee.php');
 
-$annee = new Annee();
+$controller = new \Controllers\Annee();
+$controller->lireListe();
 
-if (!isset($_SESSION["user"]) && !($_SESSION["user"]["statut"] == "Admin")) {
-
-    redirect('../../../../formulaires/formConnexion.php', '');
-} else {
-
-    $result = $annee->panorama();
-}
-?>
-<link rel="stylesheet" href="/style.css">
-<?php
-$pageTitle = "- GESTION ANNÉE";
-render('../../../', 'annee/annee', compact('pageTitle', 'result'));
-?>
