+++ dashboard/crud/annee/anneeSupprimer.php
@@ -1,35 +1,7 @@
 <?php
 session_start();
 
-require_once('../../../librairies/database/database.php');
-require_once('../../../librairies/patron.php');
-require_once('../../../librairies/models/Annee.php');
+require_once('../../../librairies/controllers/Annee.php');
 
-$annee = new Annee();
-
-
-if (!isset($_SESSION["user"]) && !($_SESSION["user"]["statut"] == "Admin")) {
-
-    redirect('../../../../formulaires/formConnexion.php', '');
-}
-
-if (isset($_GET['id']) && !empty($_GET['id'])) {
-
-    $id = strip_tags($_GET['id']);
-
-    $recuperer = $annee->supprime1($id);
-
-    if (!$recuperer) {
-        $_SESSION['erreur'] = "Cet id n'existe pas";
-        header('Location: annee.php');
-        die();
-    }
-
-    $annee->supprime2($id);
-
-    $_SESSION['message'] = "Prédiction supprimée";
-    header('Location: annee.php');
-} else {
-    $_SESSION['erreur'] = "URL invalide";
-    header('Location: annee.php');
-}
+$controller = new \Controllers\Annee();
+$controller->effacer();
