+++ librairies/controllers/Controller.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Controllers;
+
+abstract class Controller
+{
+    protected $model;
+    protected $modelName;
+
+    public function __construct()
+    {
+        $this->model = new $this->modelName();
+    }
+}
