+++ librairies/models/Annee.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Models;
+
 require_once('../../../librairies/models/Model.php');
 
 class Annee extends Model
